St. Petersburg Paradox
The St. Petersburg game is a simple, coin-tossing game that first appeared in 1738 in a
memoir by Daniel Bernoulli.   
The game ends when the coin shows heads for the first time. Formally, a single trial in 
the St. Petersburg game consists of tossing a true coin until it falls as heads. 
If it falls as heads at the nth throw,the gambler receives 2^𝑛 coins.  
(from The St. Petersburg paradox: an experimental solution by Sergio Da Silva and Raul Matsushita).
In this class you can see how an average is completely useless in a power law context. I generated 4 simulations with 1000 trials each one.
The plot of each simulation shows uttlerly different behaviours of the mean of earnings in the St. Petersburg Game.